requirements.txt
# Core FastAPI and web framework
fastapi~=0.111.0
uvicorn[standard]~=0.29.0
pydantic~=2.7.4
python-multipart~=0.0.9

# Streamlit for dashboard
streamlit~=1.36.0
plotly~=5.22.0
altair~=5.3.0
folium~=0.17.0

# Data processing and ML
pandas~=2.2.3
numpy~=1.26.4
scikit-learn~=1.5.0
scipy~=1.13.0
seaborn~=0.13.2

# Epidemic modeling - REMOVED DUE TO CONFLICTS

# AI and LLM integration
ollama~=0.1.7
python-dotenv~=1.0.0
transformers~=4.41.0
torch~=2.3.0
sentence-transformers~=2.2.2
openai~=1.35.13
anthropic~=0.23.0

# Vector database and storage
chromadb~=0.4.24
redis~=5.0.1
aioredis~=2.0.1

# Speech processing
openai-whisper==20231117 
SpeechRecognition~=3.10.0
pydub~=0.25.1

# Web scraping and APIs
requests~=2.32.3
aiohttp~=3.9.5
beautifulsoup4~=4.12.3
httpx~=0.27.0

# Security and encryption
cryptography~=42.0.8
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4

# Background tasks and async
celery~=5.4.0
asyncio-mqtt~=0.14.0

# System monitoring
psutil~=5.9.8

# Testing
pytest~=8.2.2
pytest-asyncio~=0.23.6

# Development tools
black~=24.4.2
flake8~=7.0.0
mypy~=1.10.0

# Database - Async SQLAlchemy and PostgreSQL
sqlalchemy[asyncio]~=2.0.23
asyncpg~=0.29.0
alembic~=1.13.1

# Deployment
gunicorn~=22.0.0
docker~=7.1.0